requirements.txt
streamlit
langchain-groq
langchain-community
sqlalchemy
pandas
python-dotenv
